GUI Done almost, Need to Change Project.Property
Plan is,
User Scans COM Ports,
Gets List of ports,
Selects the one which is Arduino COM Port,
Selects the correct Arduino from the List

Upon selecting the Arduino, the project SHOULD change the post-build
proeprty of the Project and add a single line in it which initiates
AVRDude command with the right parameters.

Assuming AVRDUDE is in C drive, will need to be somewhere relative to
the plugin later.